Rename votes to stakes in ValidatorProfile module
