+++ src/modules/pos/validator/components/ValidatorProfile/StakersTableHeader.js
