+++ src/modules/pos/validator/components/ValidatorProfile/ValidatorStakesView.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { useFilter } from 'src/modules/common/hooks';
-import ValidatorVotesView from './ValidatorVotesView';
+import ValidatorStakesView from './ValidatorStakesView';
 import { useReceivedVotes } from '../../hooks/queries';
 import { mockReceivedVotes } from '../../__fixtures__';
 
@@ -11,7 +11,7 @@ const mockClearFilters = jest.fn();
 jest.mock('../../hooks/queries');
 jest.mock('src/modules/common/hooks');
 
-describe('Validator votes view', () => {
+describe('Validator stakes view', () => {
   useReceivedVotes.mockReturnValue({ data: mockReceivedVotes });
 
   const props = {
@@ -24,23 +24,23 @@ describe('Validator votes view', () => {
     clearFilters: mockClearFilters,
   });
 
-  it('Should render a list of voters', () => {
-    render(<ValidatorVotesView {...props} />);
+  it('Should render a list of stakers', () => {
+    render(<ValidatorStakesView {...props} />);
 
-    expect(screen.getByText('Voters')).toBeTruthy();
+    expect(screen.getByText('Stakers')).toBeTruthy();
     mockReceivedVotes.data.votes.forEach(({ name }) => {
       expect(screen.getByText(name)).toBeTruthy();
     });
   });
 
-  it('Should filter voters by search value', () => {
+  it('Should filter stakers by search value', () => {
     useFilter.mockReturnValue({
       filters: { adress: props.address },
       applyFilters: mockApplyFilters,
       clearFilters: mockClearFilters,
     });
 
-    render(<ValidatorVotesView {...props} />);
+    render(<ValidatorStakesView {...props} />);
 
     const searchField = screen.getByTestId('addressFilter');
     fireEvent.change(searchField, { target: { value: 'test' } });
@@ -57,7 +57,7 @@ describe('Validator votes view', () => {
     });
     useReceivedVotes.mockReturnValue({ data: {} });
 
-    render(<ValidatorVotesView {...props} />);
+    render(<ValidatorStakesView {...props} />);
 
     expect(screen.queryByTestId('addressFilter')).toBeFalsy();
     expect(screen.getByText('(...)')).toBeTruthy();
